llm-tests/api/requirements.txt
bardapi
python-gemini-api
aiohttp
pycookiecheat
pandas
nltk
bs4
openpyxl
fastapi
uvicorn[standard]
sqlalchemy
databases
psycopg2-binary
alembic
asyncpg
redis==4.3.5